requirements.txt
tensorflow
tensorflow-model-optimization
flask
pillow
numpy
psutil
